Move TextCollector logic to TextCollectorTransformer
